+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Use positional parameters when appropriate
 * Drop logging class down to global methods
+* Make basil a global executable
 
 ## 1.0.10